created cat_request_handler to store modular code used by create and update views. created and tested update view
